Committing the treediff for the changesdue scenario
NOTES:
- during a commit in the commerce scenario, the ice specific changes were committed to Xaraya Classic. On committing the
commerce scenario into the monotone branch, these changes were left out (they will return in the ice theme).
As the changesdue scenario had merged in the changes of the commerce scenario, the same changes to Xaraya Classic were
also omitted here. <-- Probably only relevant to Marc! 

Note that a treediff is a somewhat simplistic way to commit a whole scenario. 
Changes will get registered properly, but keep an eye on the following situations when testing:
  
- if the scenario REMOVED a file, it will reappear in this repository (i check this manually, but i suck at that)
- if the host repository ADDS a file and the scenario REMOVES that again, it will reappear
- there might be more complex situation where the diff is not quite up to par, in general, CHECK YOUR SCENARIO