Crazy problem:
Apache segfaulted for no apparent reason leading to all kinds of weird stuff (most visibly the DD xarinit.php function running twice!!)

As it turned out, the E_STRICT errorlevel (which is a pita anyway) caused many errors (like more than a 1000 in one run) that it probably flooded the php5 lib (or apache) with some kind of buffer overflow or range exception (guessing here though)

This was observed on PHP 5.1.2 and 5.1.4 on debian/linux with apache 2.0.55

Rolling back the code to before a revision which fixed the E_STRICT issue (so, rolling back to a rev which contained a bug in the error handling due to the E_STRICT not being included in E_ALL) reveiled a hint.

Effectively, since E_STRICT errors mostly come in floods, makes E_STRICT almost unusable to enable. Assuming my environment is somehow special (others could not reproduce this error) I've still left E_STRICT + E_ALL to be the default in 2.x when nothing is explicitly specified in config.system.php

	* html/includes/exceptions/handlers.php: 
	    document and clarify what goes on during the handling of a PHP error
	* html/modules/dynamicdata/class/datastores.php: 
	    these return by references were causing the flood