A prototype example of implementing a protocol handler for the more and more common stream interface in PHP. This example was copied almost verbatim from the example given at php.net, with the adaptations needed to make it actually do something.
Committing the example as i think 'protocol' wrapping and handling can be of 
great interest to the core. The example itself should at some point be dropped
as it is as such of no practical use (but who knows).

The stream thought was triggered by several thoughts on (memory) templates in BL,
stream access to DD, stream access to content and other time-varying data in 
xaraya.

	* html/includes/streams: new dir
	* html/includes/streams/variables.php: new file