XARAYA.RC change: define a validate_commit_message hook for 2 purposes: 	1. Being able to validate a commit message (one of length two could be wrong for example) 	2. The rcfile is taking into account that there may exist a directory with QA tests floating around, new mtn-lua style
Using this rc file without those tests present does nothing different (but checking if the tests are available). 
If they are, mtn commit uses them and refuses to commit is one of the tests fail.

The actual tests are coming in thru a commit on com.xaraya.qa.tests later on.

	* tools/monotone/xaraya.rc: ...