Eval free compilation forever :-)
Seriously,
- we now *require* a temporary directory (tpl cache dir for now) even when the
template cache-ing if set to off. Setting it to off means we reuse a temporary 
file for each compilation action in the tpl cache dir. This eliminate eval completely 
and is compatible with isp's where as the non-eval solution we had used tempname() 
which is typically in /tmp, so in safe_mode that would not have worked anyway
- because of this we can simplify a bit, the xarTpl__execute function is gone 
since we dont need to do that anymore.

Should we decide that the creation of that one file in the tpl cache is not wanted
we should consider (we should anyways) making the input to the compiler stream
based, so we can stream xar templates from anywhere. The current classed should
be a good starting point, but we dont actually need it atm.

This completes the first refactoring for xarTemplate in combo with compilation.
I think the cache is next, or perhaps some more function OO-ing. we'll see.
	* html/includes/xarTemplate.php: ...
	* html/modules/roles/xaradmin/sendmail.php: ...
	* html/modules/roles/xartemplates/admin-displayrole.xd: ...