Introduce the 'Object' class of which all other classes will be derived.
As of now, this class is empty, to keep all options open. It's not trivial to
add methods and/or properties to this class without introducing possibly unwanted
restrictions. The 'one root class' paradigm is attractive enough to warrant the 
empty class, while keeping a place in the code where *if* we need something for 
all classes we can put it in.

Now, this change is small, but the work generated by it is lots. We have to go over
all classes in the codebase which we define and if a class is defined like:

class MyClass 
{
....
}

it has to change into

class MyClass extends Object 
{
....
}

This can be done by everyone having a spare moment to go over files. It has no
effect on functionality, but needs to be done code-wide. 

Other tasks, but less trivial than the above, which come out of this are:
	- when there are parameters to methods in classes which take some sort of
	  object, a so called 'type hint' should be added now, i.e.:
	  	function someMethod( $someObject )
	  becomes
		function someMethod( Object $someObject)	// $someObject must be of class 'Object' or one of its descendents
	This enables type checking during 'compile time' vs 'runtime' so we can generate
	warnings or errors early.

	- Unrelated but similar are: 
		function someMethod( Array $someArray))         // type hinting on $someArray to be an array type variable
		function someMethod( SomeInterface $someObject) // type hinting on $someObject to have implemented the interface 'SomeInterface'
		
	- possibly some checks on is_object() will become redundant


	* html/includes/xarPreCore.php: ...