Insert a @todo in the handlers class on a change we need here.
I noticed that the trace in exceptions was gone, the change that introduced
that, also introduced a new public method in the handler class. The idea behind
the handlers class was that the class was suitable for chaining with handler
fallback following the inheritance. Apparently the current design does not \
encourage that, and instead makes chaining methods more easy. We should change
that. 

I'm rambling on this because this will be essential in our 'framework' paradigm 
where we can mix and match formats at will. This requires a strict exception 
chain we can trust completely.

The change is not hard, but not trivial either. This lengthy comment will 
attract enough attention to make me do this change.