Mass changing *.xd -> *.xt
Having both xd and xt extension serves no purpose, sync them up to xt which 
makes the most sense.

This revision was committed without using the QA checks as I did not want to 
mix content changes with the actual xd to xt change.

The commit is on the xd2xt branch which gives us an opportunity to play with
propagates on this change, which are likely to be problematic.

If they are, we should ignore this branch and introduce a flag day to perform 
the steps in this revision, which are:
    1. do the patches in the handlers and template, which are neigh trivial
    2. the command i used on osx for the name changes was:
    find . -name '*.xd' | sed -e 's/\.xd// | xargs -n1 -I % mtn mv %.xd %.xt
    [the -I option is different from the linux one, which uses -i. I do not 
    know how/if windows handles xargs]
    
If the propagates go fine, people can merge whenever they are ready.